+++ web/modules/custom/hoeringsportal_audit_log/src/EventSubscriber/ControllerListener.php
@@ -40,10 +40,19 @@ final class ControllerListener implements EventSubscriberInterface {
    */
   public function onController(ControllerEvent $event): void {
     $pathInfo = $event->getRequest()->getPathInfo();
+    $requestUri = $event->getRequest()->getRequestUri();
+    $urlPatterns = $this->configHelper->getUrlPatterns();
+
+    foreach ($urlPatterns as $urlPattern) {
+      if (@preg_match($urlPattern, $requestUri)) {
+        $this->logAuditMessage($pathInfo);
+        return;
+      }
+    }
+
     $currentRouteName = $event->getRequest()->attributes->get('_route');
     $routesThatShouldBeLogged = $this->configHelper->getRouteNames();
-
-    if (in_array($currentRouteName, $routesThatShouldBeLogged)) {
+    if ($routesThatShouldBeLogged && in_array($currentRouteName, $routesThatShouldBeLogged)) {
       // Early return, if the route is in config no need to do anything besides
       // auditlog.
       $this->logAuditMessage($pathInfo);